# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/getwallet": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get wallet history of user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User.Payment"
                ],
                "summary": "Get wallet history",
                "operationId": "GetWallet",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WalletRes"
                        }
                    },
                    "400": {
                        "description": "can't get product"
                    }
                }
            }
        },
...
        "/user/order/checkout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "place order",
                "tags": [
                    "User.Order"
                ],
                "summary": "API for order cart products",
                "operationId": "OrderCartProducts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "coupon",
                        "name": "coupon",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pay method",
                        "name": "payMethod",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "html added",
                        "schema": {
                            "$ref": "#/definitions/response.Order"
                        }
                    },
                    "400": {
                        "description": "somthing wrong!!"
                    },
                    "406": {
                        "description": "string \"Invalid input"
                    }
                }
            }
        },
...
        "/user/order/payment/addmonytowallet": {
            "get": {
                "description": "Make payment using razorpay page .",
                "tags": [
                    "User.Payment"
                ],
                "summary": "Make payment razorpay",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Amount",
                        "name": "Amount",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "faild"
                    }
                }
            }
        },
        "/user/order/payment/checkout/razorpay": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User.Payment"
                ],
                "summary": "api for user pay section",
                "operationId": "RazorPayCheckout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "coupon",
                        "name": "coupon",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "html added successfully"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "404": {
                        "description": "invalid "
                    }
                }
            }
        },
        "/user/order/payment/razorpay/process-order": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Verify razorpay payment using razorpay credentials .",
                "tags": [
                    "User"
                ],
                "summary": "api for user pay section",
                "operationId": "ProccessRazorpayOrder",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "faild"
                    },
                    "406": {
                        "description": "invalid input"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Order": {
            "type": "object",
            "properties": {
                "Razor_pay_order_id": {
                    "type": "string"
                },
                "actual_price": {
                    "type": "number"
                },
                "disc_price": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "net_amount": {
                    "type": "number"
                },
                "order_date": {
                    "type": "string"
                },
                "order_status_id": {
                    "type": "integer"
                },
                "payment_id": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "users_id": {
                    "type": "integer"
                }
            }
        },
...
        "response.WHistoty": {
            "type": "object",
            "properties": {
                "added_at": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "razorpay_payment_id": {
                    "type": "string"
                }
            }
        },
        "response.WalletRes": {
            "type": "object",
            "properties": {
                "balence": {
                    "type": "number"
                },
                "historys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.WHistoty"
                    }
                }
            }
        }
    }
}